feat: reorganized cache functions and logic
